Add example type checking annotation
This is a demo that adds a type checker. For now it only supports types
"Object" and "Optional[Object]", though it could be extended to some
simple types like Tuples, Lists, Dicts, etc which contain Objects.

The idea is that you can set DRGN_TYPE_CHECK=1, and any function with
the given annotation is type checked at runtime. There's no runtime cost
when the annotation is disabled, so it's safe to leave in your code.

The key trade-off made here is that we expect the C type annotation to
be in the docstring. This is simply because my code convention is that
every docstring for an Object needs a C type description, inside the
double-backtick for readability on Sphinx docs. It's quite simply my
preference. There is another option, typing.Annotated, which allows you
to include additional metadata for type checks. This would allow the C
type information to live with the Python types, but the downside to that
is that it probably wouldn't look nice in Sphinx (if it gets displayed
at all). It doesn't seem hard to implement support for that instead.

The actual C type checks are just string comparison for type_name(). I
suppose I should at least tokenize it so that it's not whitespace
sensitive, but that hasn't been an issue in my testing.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>