libdrgn: vector: support storing entries in vector structure
Add an inline_size parameter to DEFINE_VECTOR()/DEFINE_VECTOR_TYPE()
which specifies how many entries should be stored inline in the struct
vector. This is similar to the std::string small string optimization
(SSO) [1] and LLVM's SmallVector [2]. It allows avoiding malloc() and a
cache miss for small vectors at the cost of an extra couple of branches.
vector_steal() is also undefined for vector types with inline entries.

1: https://stackoverflow.com/questions/10315041/meaning-of-acronym-sso-in-the-context-of-stdstring/10319672#10319672
2: https://llvm.org/doxygen/classllvm_1_1SmallVector.html

Signed-off-by: Omar Sandoval <osandov@osandov.com>