libdrgn: dwarf_info: scalably index all DIEs per name
We currently deduplicate entries in the DWARF index by (name, tag, file
name). We want to add support for looking up nested classes, so this is
a problem: not every DIE defining a class also defines all of its nested
types, so the one DIE we index may not allow us to find every nested
class. Instead, we need to index every DIE with a given name.

This sounds horribly expensive, both in terms of CPU and memory, but we
can mitigate this in several ways:

- We no longer need to parse the file name table, cache file name
  hashes, parse DW_AT_decl_file, or store the file name hash for indexed
  DIEs.
- Instead of storing the tag for each indexed DIE, we can split the DIE
  map into a map per tag.
- We can store the DIEs matching a name in a vector instead of a linked
  list.
- We can use the new inline entry and small size variants of vectors.
- We can move struct drgn_namespace_dwarf_index * to a tree separate
  from the indexed DIEs. After all of these changes, we only need a
  single uintptr_t per indexed DIE.
- We can get rid of the struct drgn_dwarf_index_pending_die list for a
  namespace and use the indexed DIEs instead, which are half the size.
- DW_TAG_base_type maps can be assumed to be globally unique, so they
  can be stored in their own map of one DIE indexed only by name.
- Each thread can independently build the DIE maps without any
  synchronization to be merged at the end.

Here are some performance results comparing the New version (this
commit) to the Old version (commit 16164dbe6edc ("libdrgn: detect
flattened vmcores and raise error")). Application is either a large,
statically-linked C++ application or the live Linux kernel. Threads is
the OMP_NUM_THREADS setting used (the machine used for testing has 80
CPUs). Time is the amount of time it took to load and index debugging
information. Anon is the amount of anonymous (e.g., heap) memory used.
File is the amount of file memory used. Large C++

Application | Threads | Version | Time   | Anon   | File
------------+---------+---------+--------+--------+-------
Large C++   | 80      | New     | 5 s    | 3.5 GB | 1.4 GB
            |         | Old     | 15 s   | 5.2 GB | 1.7 GB
            | 8       | New     | 6.5 s  | 3.4 GB | 1.4 GB
            |         | Old     | 10 s   | 5.2 GB | 1.7 GB
            | 1       | New     | 30 s   | 3.4 GB | 1.4 GB
            |         | Old     | 51 s   | 5.2 GB | 1.7 GB
Linux       | 80      | New     | 270 ms | 128 MB | 300 MB
            |         | Old     | 380 ms | 73 MB  | 326 MB
            | 8       | New     | 240 ms | 115 MB | 300 MB
            |         | Old     | 240 ms | 73 MB  | 326 MB
            | 1       | New     | 700 ms | 87 MB  | 300 MB
            |         | Old     | 800 ms | 73 MB  | 326 MB

The results show that the new approach is almost always faster. For the
large C++ application, it is much better for both time and memory usage.
For the Linux kernel, it is slightly faster and uses more anonymous
memory, although that is partially offset by less file memory. (For the
Linux kernel, there is a dip in performance for both approaches from 8
threads to 80 which is worth looking into later.)

Signed-off-by: Omar Sandoval <osandov@osandov.com>