Allow Program.set_core_dump() and program_from_core_dump() to accept FD
This allows an open file descriptor to be passed into Drgn and treated
as a core dump.  There are many use cases, but one interesting one is
that the FD could be sent by a helper process running as root, allowing
a non-root Drgn process to debug the running kernel.

Signed-off-by: Stephen Brennan <stephen@brennan.io>