drgn.helpers.linux.printk: handle log_buf name conflict between v3.18 and v4.15
In a few kernel versions with structured printk, printk and the BPF
verifier both had static global variables named log_buf.
get_printk_records() doesn't handle this. We never hit this before with
the old DWARF index, but the rework exposed this for some reason. Pass
the filename so we always get the right one.

Fixes: ba51e8840b8e ("helpers: add get_printk_records()")
Signed-off-by: Omar Sandoval <osandov@osandov.com>