drgn.cli: look up symbols as missing in global namespace
Introduce a new option --auto-find to make the interactive mode load
the corresponding symbol from prog if the user code access a global
variable that is undefined while the name is defined by the program
being debugged.

For example, you can access jiffies of the kernel directly without
going through prog['jiffies'].

  print(jiffies)

This line will be compiled and run. However, the auto-find mode will
look into the code object and know that the code object is going to
access a global variable named "jiffies".  However, "jiffies" is not
defined in the global namespace.  The auto-find mode will check if the
program being debugged defines a symbol with the same name and add it
to the global namespace if there is.

Signed-off-by: Thinker Lee <thinker.li@gmail.com>