helpers: Clarify type requirements for bitops
The docs currently state that the bitmap must be an "unsigned long *".
This is similar to the kernel API. It may not be immediately obvious
"unsigned long[]" arrays are acceptable as well, so clarify that.

Also, update the module documentation to reflect the fact that on
little-endian architectures, the helpers work just as well with any
other unsigned integer type, and update the tests to verify this.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>