libdrgn: stack_trace: get rid of PRSTATUS PID match heuristic
Commit 1b47b866b440 ("libdrgn: go back to trusting PRSTATUS PID")
introduced a check so that we only use the PRSTATUS note for stack
unwinding if its PID field matches the PID of the task we're trying to
unwind. This was intended to detect when a CPU was in the middle of a
context switch during a crash. However, it has caused more trouble than
it's worth. For example:

1. It's broken on s390x, which puts the CPU number + 1 in the PID field.
   See also commit 7cb3e99b2389 ("libdrgn: program: find crashed task
   with cpu_curr() instead of find_task()").
2. It's broken on dumps from QEMU's dump-guest-memory command, which
   also puts the CPU number + 1 in the PID field. See issue #356.
3. It breaks the reasonable assumption that
   prog.stack_trace(cpu_curr(prog, cpu)) always gets the stack trace for
   a given CPU. This slowed down an internal scheduler investigation.

It doesn't even really accomplish what it's trying to do, since the
check itself is racy. Let's remove it (which is not as simple as it
sounds, since we were also using it in lieu of a proper on_cpu fallback
for !SMP kernels).

Signed-off-by: Omar Sandoval <osandov@osandov.com>