Revert "libdrgn: add last virtual address hint to page table iterator"
This reverts commit 747e02857d9009e1ff93039cc838d636e95900d3 (except for
the test improvements). Peter Collingbourne noticed that the change I
used to test the performance of reading a single PTE at a time [1]
didn't cache higher level entries. Keeping that caching makes the
regression I was worried about negligible. So, there's no reason to add
the extra complexity of the hint.

1: https://github.com/osandov/drgn/pull/312#issuecomment-1754082129

Signed-off-by: Omar Sandoval <osandov@osandov.com>