cli: Add a command system
As a debugger library, drgn can be a little bit verbose when running
interactively. There has been general concensus that it would be nice to
have a way to define shorthands for common operations, or define user
scripts that are easy to run within the shell. However we have yet to
settle upon a solution. Implementing a full CLI seems out of the scope
of drgn. Leveraging the power of a system like IPython's "magic"
commands is promising, but adds heavy dependencies.

One example CLI to draw inspiration from is SQLite, which uses the "."
prefix to define meta-commands that aren't part of the SQL language.
Python, like SQL, cannot begin a statement with a ".", so it could be
used to unambiguously signal that a command is being entered, and thanks
to the code.InteractiveConsole, we can easily extend the console to
handle these commands.

So, let's add a command system. Commands are simply callables which take
a program, the full line of text (which may contain arguments), and the
CLI locals. They may return a Python object which is stored to
`builtins._`. They can be registered using a decorator, and an API is
available to retrieve the default list. The run_interactive() function
allows the user to customize the available command list.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>