docs: fix setuptools_config to not force full rebuild every time
Every build_sphinx command does a full rebuild with the following log:

  updating environment: [config changed ('project')] 14 added, 0 changed, 0 removed

It looks like Sphinx is comparing the config saved at the end of a run
to the config before the builder-inited event. Fix it by connecting to
the config-inited event first, which according to the Sphinx
documentation [1] happens earlier.

1: https://www.sphinx-doc.org/en/master/extdev/appapi.html#sphinx-core-events

Signed-off-by: Omar Sandoval <osandov@osandov.com>