CONTRIBUTING: drop support for developing on EOL Python versions
Since updating pre-commit and the hooks, development on EOL Python
versions is no longer supported, and will fail. Update the contributing
guide to explain that Python 3.6 is supported for build, runtime, and
tests, but development workflow scripts require a supported Python
version. Repeat this warning in the pre-commit section, where it is most
relevant.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>