x86_64: mm: support numa migration pte entry
The PRESENT bit alone does not indicate whether a page actually exists. The
PRESENT bit is also cleared for numa migration pte entries, even though the
page is present. Therefore, to determine if a page truly exists, we need to
check pte and convert it to phys in the same way the kernel does.

Signed-off-by: Hao Li <haoli.tcs@gmail.com>