docs: Prevent jarring 50% width for medium screens
For screens that aren't terribly wide, there's a break point where the
current documentation design will snap to 50% width, and the text is
difficult to read due to the narrow layout. Rather than using 50% as the
rule, let's setup a couple good breakpoints that are wider than
Alabaster's default.

In particular, at the 875px breakpoint the sidebar appears, removing
220px from the content. So add a breakpoint at 875 + 220 px, allowing
the content width to grow up until that point, at which point we can cap
the width.

Fixes: 98cf8d51 ("docs: make page body wider")

Signed-off-by: Stephen Brennan <stephen@brennan.io>