kallsyms: add symbol finder for live & coredump
The Linux kernel can be configured to include kallsyms, a built-in
compressed symbol table which is also exposed at /proc/kallsyms. The
symbol table contains most (but not all) of the ELF symbol table
information. It can be used as a Symbol finder.

The kallsyms information can be extracted in two ways: for live systems
where we have root access, the simplest approach is to simply read
/proc/kallsyms. For vmcores, or live systems where we are not root, we
must parse the data from the vmcore, which is significantly more
involved.

To avoid tying the kallsyms system too deeply into the drgn internals,
the finder is exposed as a Python class, which must be created using
symbol information from the vmcoreinfo. Attaching the KallsymsFinder to
the program will attach the underlying C function, so we can avoid some
of the inefficiencies of the Python API.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>