helpers.linux.percpu: add helper to get owner cpu of a per-cpu variable.
Sometimes it is useful to get cpu that owns a given per-cpu variable.
For example osq locks and qspinlocks use per-cpu entities when a CPU
needs to spin, while waiting to acquire the lock. This happens when
there are multiple spinners for a given lock.
In such cases, this helper can let us know the spinners for a given
qspinlock or osq lock.

Signed-off-by: Imran Khan <imran.f.khan@oracle.com>