contrib/btrfs_tree.py: flesh out structures and helpers
Add support for parsing and printing most of the Btrfs item types and
clean up/add some tree search and extent buffer helpers. I used this to
debug a Btrfs crash in production. It's almost ready to be turned into
proper helpers, but it needs lots of test cases and a big more cleanup.

Signed-off-by: Omar Sandoval <osandov@osandov.com>