packit: disable EPEL 9 builds
drgn is now in RHEL 9.4, so it was retired from EPEL 9. The EPEL 9
Packit build is failing with

  No matching package to install: 'libkdumpfile-devel'

It might still be useful to build for RHEL 9, but Davide Cavalca is
looking into what the best thing to do here is. For now, disable the
EPEL 9 builds to get the CI green.

Signed-off-by: Omar Sandoval <osandov@osandov.com>