drgn.helpers.linux.mm: update PageSlab() for Linux 6.10
Linux 6.10 switched PageSlab() from using a page flag to a page type.
This is annoying since page types are all represented using macros
instead of enums, but there's a VMCOREINFO entry we can use. A better
solution is to change the page type definitions to enums in the kernel
source. We're looking into that upstream, but for now let's get the
tests passing on 6.10.

Signed-off-by: Omar Sandoval <osandov@osandov.com>