drgn.helpers.common.memory: add optional cache to identify_address()
print_annotated_memory() and print_annotated_stack() are pretty slow
when called on large memory regions, largely because identify_address()
walks every task for every call on a non-direct map address in order to
resolve vmap stacks. We can cache the mapping from a vmap stack area to
a task, but we need callers to be able to indicate when it's okay to
cache in order to avoid stale results. To do so, add a cache parameter
to identify_address(), which is just an opaque dictionary that should be
passed between successive calls.

Signed-off-by: Omar Sandoval <osandov@osandov.com>